# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Xito Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Client {
  id: number;
  name: string;
  email: string;
  phone: string;
  /** active | inactive | completed */
  status: string;
  /** @nullable */
  weddingDate?: string | null;
  /** @nullable */
  eventLocation?: string | null;
  /** @nullable */
  packageType?: string | null;
  /** @nullable */
  totalAmount?: number | null;
  /** @nullable */
  advancePaid?: number | null;
  /** @nullable */
  remainingBalance?: number | null;
  /**
   * pending | partial | paid
   * @nullable
   */
  paymentStatus?: string | null;
  /** @nullable */
  notes?: string | null;
  tags?: string[];
  /**
   * 0-100 percentage
   * @nullable
   */
  progress?: number | null;
  /** @nullable */
  referredBy?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ClientInput {
  name: string;
  email: string;
  phone: string;
  status?: string;
  /** @nullable */
  weddingDate?: string | null;
  /** @nullable */
  eventLocation?: string | null;
  /** @nullable */
  packageType?: string | null;
  /** @nullable */
  totalAmount?: number | null;
  /** @nullable */
  advancePaid?: number | null;
  /** @nullable */
  remainingBalance?: number | null;
  /** @nullable */
  paymentStatus?: string | null;
  /** @nullable */
  notes?: string | null;
  tags?: string[];
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  referredBy?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface ClientUpdate {
  name?: string;
  email?: string;
  phone?: string;
  status?: string;
  /** @nullable */
  weddingDate?: string | null;
  /** @nullable */
  eventLocation?: string | null;
  /** @nullable */
  packageType?: string | null;
  /** @nullable */
  totalAmount?: number | null;
  /** @nullable */
  advancePaid?: number | null;
  /** @nullable */
  remainingBalance?: number | null;
  /** @nullable */
  paymentStatus?: string | null;
  /** @nullable */
  notes?: string | null;
  tags?: string[];
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  referredBy?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface Project {
  id: number;
  name: string;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  /** Booked | Editing | Preview Sent | Final Delivery | Completed */
  status: string;
  /**
   * Photography | Videography | Photo+Video | Album | Highlights
   * @nullable
   */
  type?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  deliveryDate?: string | null;
  /** @nullable */
  deliveryLink?: string | null;
  /** @nullable */
  notes?: string | null;
  /**
   * low | medium | high
   * @nullable
   */
  priority?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  tags?: string[];
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ProjectInput {
  name: string;
  clientId: number;
  status: string;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  deliveryDate?: string | null;
  /** @nullable */
  deliveryLink?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  tags?: string[];
}

export interface ProjectUpdate {
  name?: string;
  clientId?: number;
  status?: string;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  deliveryDate?: string | null;
  /** @nullable */
  deliveryLink?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  tags?: string[];
}

export interface ProjectStatusUpdate {
  status: string;
}

export interface FileRecord {
  id: number;
  name: string;
  originalName?: string;
  /** photo | video | document | other */
  type: string;
  /** @nullable */
  size?: number | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  deliveryLink?: string | null;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  projectName?: string | null;
  /** @nullable */
  uploadedBy?: string | null;
  /** available | processing | archived | deleted */
  availability?: string;
  /** none | pending | backed_up */
  backupStatus?: string;
  downloadCount?: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface FileInput {
  name: string;
  originalName?: string;
  type: string;
  /** @nullable */
  size?: number | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  deliveryLink?: string | null;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  uploadedBy?: string | null;
  availability?: string;
  backupStatus?: string;
  /** @nullable */
  notes?: string | null;
}

export interface FileUpdate {
  name?: string;
  /** @nullable */
  deliveryLink?: string | null;
  availability?: string;
  backupStatus?: string;
  /** @nullable */
  notes?: string | null;
}

export interface Notification {
  id: number;
  /** payment_due | deadline | wedding_upcoming | delivery_pending | general */
  type: string;
  title?: string;
  message: string;
  isRead: boolean;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  totalClients: number;
  activeProjects: number;
  pendingDeliveries: number;
  completedWeddings: number;
  totalRevenue: number;
  collectedRevenue: number;
  pendingRevenue: number;
  thisMonthBookings: number;
  /** @nullable */
  storageUsed?: number | null;
  totalFiles?: number;
}

export interface MonthlyAnalytics {
  month: number;
  year: number;
  monthName?: string;
  bookings: number;
  revenue: number;
  projectsCompleted: number;
}

export interface UpcomingEvent {
  id: number;
  /** wedding | deadline | delivery | payment */
  type: string;
  title: string;
  date: string;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  projectName?: string | null;
  daysUntil?: number;
  /** low | medium | high | critical */
  urgency?: string;
}

export interface StatusCount {
  status: string;
  count: number;
  percentage?: number;
}

export interface Tag {
  id: number;
  name: string;
  /** @nullable */
  color?: string | null;
  createdAt?: string;
}

export interface TagInput {
  name: string;
  /** @nullable */
  color?: string | null;
}

export type ListClientsParams = {
  search?: string;
  status?: string;
  tag?: string;
};

export type ListProjectsParams = {
  clientId?: number;
  status?: string;
  search?: string;
};

export type ListFilesParams = {
  projectId?: number;
  clientId?: number;
  type?: string;
};

export type ListNotificationsParams = {
  unreadOnly?: boolean;
};

export type GetMonthlyAnalyticsParams = {
  year?: number;
};