# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Xito Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── PROJECTS ──────────────────────────────────────────────────────────────────
...
  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files
      parameters:
        - name: projectId
          in: query
          schema: { type: integer }
        - name: clientId
          in: query
          schema: { type: integer }
        - name: type
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileRecord"
    post:
      operationId: createFile
      tags: [files]
      summary: Create file record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRecord"
  /files/{id}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get file by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: File details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRecord"
    patch:
      operationId: updateFile
      tags: [files]
      summary: Update file record
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRecord"
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete file record
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── NOTIFICATIONS ──────────────────────────────────────────────────────────
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, name, email, phone, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          description: "active | inactive | completed"
        weddingDate:
          type: ["string", "null"]
        eventLocation:
          type: ["string", "null"]
        packageType:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
        advancePaid:
          type: ["number", "null"]
        remainingBalance:
          type: ["number", "null"]
        paymentStatus:
          type: ["string", "null"]
          description: "pending | partial | paid"
        notes:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        progress:
          type: ["integer", "null"]
          description: "0-100 percentage"
        referredBy:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    ClientInput:
      type: object
      required: [name, email, phone]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
        weddingDate:
          type: ["string", "null"]
        eventLocation:
          type: ["string", "null"]
        packageType:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
        advancePaid:
          type: ["number", "null"]
        remainingBalance:
          type: ["number", "null"]
        paymentStatus:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        progress:
          type: ["integer", "null"]
        referredBy:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
        weddingDate:
          type: ["string", "null"]
        eventLocation:
          type: ["string", "null"]
        packageType:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
        advancePaid:
          type: ["number", "null"]
        remainingBalance:
          type: ["number", "null"]
        paymentStatus:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        progress:
          type: ["integer", "null"]
        referredBy:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]

    # ── PROJECT SCHEMAS ───────────────────────────────────────────────────────
...
    FileRecord:
      type: object
      required: [id, name, type, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        originalName:
          type: string
        type:
          type: string
          description: "photo | video | document | other"
        size:
          type: ["integer", "null"]
        url:
          type: ["string", "null"]
        deliveryLink:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        projectName:
          type: ["string", "null"]
        uploadedBy:
          type: ["string", "null"]
        availability:
          type: string
          description: "available | processing | archived | deleted"
        backupStatus:
          type: string
          description: "none | pending | backed_up"
        downloadCount:
          type: integer
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    FileInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        originalName:
          type: string
        type:
          type: string
        size:
          type: ["integer", "null"]
        url:
          type: ["string", "null"]
        deliveryLink:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        uploadedBy:
          type: ["string", "null"]
        availability:
          type: string
        backupStatus:
          type: string
        notes:
          type: ["string", "null"]
    FileUpdate:
      type: object
      properties:
        name:
          type: string
        deliveryLink:
          type: ["string", "null"]
        availability:
          type: string
        backupStatus:
          type: string
        notes:
          type: ["string", "null"]

    # ── NOTIFICATION SCHEMAS ──────────────────────────────────────────────────
    Notification:
      type: object
      required: [id, type, message, isRead, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          description: "payment_due | deadline | wedding_upcoming | delivery_pending | general"
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        clientId:
          type: ["integer", "null"]
        projectId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── ANALYTICS SCHEMAS ─────────────────────────────────────────────────────
...