# lib/api-zod/src/generated/types/clientInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Xito Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ClientInput {
  name: string;
  email: string;
  phone: string;
  status?: string;
  /** @nullable */
  weddingDate?: string | null;
  /** @nullable */
  eventLocation?: string | null;
  /** @nullable */
  packageType?: string | null;
  /** @nullable */
  totalAmount?: number | null;
  /** @nullable */
  advancePaid?: number | null;
  /** @nullable */
  remainingBalance?: number | null;
  /** @nullable */
  paymentStatus?: string | null;
  /** @nullable */
  notes?: string | null;
  tags?: string[];
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  referredBy?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}