# lib/api-zod/src/generated/types/fileRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Xito Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface FileRecord {
  id: number;
  name: string;
  originalName?: string;
  /** photo | video | document | other */
  type: string;
  /** @nullable */
  size?: number | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  deliveryLink?: string | null;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  projectName?: string | null;
  /** @nullable */
  uploadedBy?: string | null;
  /** available | processing | archived | deleted */
  availability?: string;
  /** none | pending | backed_up */
  backupStatus?: string;
  downloadCount?: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}